+++ app/src/test/java/woowacourse/shopping/ui/productdetail/contract/presenter/ProductDetailPresenterTest.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.ui.productdetail.contract.presenter
 
-import com.example.domain.model.CartRepository
+import com.example.domain.repository.CartRepository
 import com.example.domain.model.Product
-import com.example.domain.model.RecentRepository
+import com.example.domain.repository.RecentRepository
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
