+++ domain/src/main/java/com/example/domain/repository/RecentRepository.kt
@@ -1,4 +1,6 @@
-package com.example.domain.model
+package com.example.domain.repository
+
+import com.example.domain.model.Product
 
 interface RecentRepository {
     fun insert(product: Product)
